Tamojit files placed in correct folder
