+++ package.json
@@ -13,9 +13,8 @@
     "react": "16.5.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-31.0.0.tar.gz",
     "react-native-elements": "^1.0.0-beta7",
-    "react-native-vector-icons": "^6.1.0"
-    "react-native-gesture-handler": "^1.0.12",
     "react-native-vector-icons": "^6.1.0",
+    "react-native-gesture-handler": "^1.0.12",
     "react-navigation": "^3.0.9",
     "react-redux": "^6.0.0",
     "redux": "^4.0.1",
