+++ src/components/FlashScreen.js
